[FEAT]최근조회,저장레시피
API 연결 완료, 상세화면 연결 완료

+++ mohaemukzip/Service/Profile/ProfileResponseDTO.swift
@@ -35,7 +35,7 @@ enum ProfileResponseDTO {
         let videoDuration: String
 
         let cookingTimeMinutes: Int?
-        let difficulty: Int?
+        let difficulty: Double?
 
         let isBookmarked: Bool
     }
